Pulls out runDemo into separate file
